requirements.txt
Django==5.2.7
djangorestframework==3.15.2
django-cors-headers==4.3.1
Pillow==10.0.1
python-decouple==3.8

